+++ actix-http/src/h1/timer.rs
@@ -1,6 +1,7 @@
 use std::{fmt, future::Future, pin::Pin, task::Context};
 
 use actix_rt::time::{Instant, Sleep};
+use tracing::trace;
 
 #[derive(Debug)]
 pub(super) enum TimerState {
@@ -24,7 +25,7 @@ impl TimerState {
 
     pub(super) fn set(&mut self, timer: Sleep, line: u32) {
         if matches!(self, Self::Disabled) {
-            log::trace!("setting disabled timer from line {}", line);
+            trace!("setting disabled timer from line {}", line);
         }
 
         *self = Self::Active {
@@ -39,11 +40,11 @@ impl TimerState {
 
     pub(super) fn clear(&mut self, line: u32) {
         if matches!(self, Self::Disabled) {
-            log::trace!("trying to clear a disabled timer from line {}", line);
+            trace!("trying to clear a disabled timer from line {}", line);
         }
 
         if matches!(self, Self::Inactive) {
-            log::trace!("trying to clear an inactive timer from line {}", line);
+            trace!("trying to clear an inactive timer from line {}", line);
         }
 
         *self = Self::Inactive;