+++ actix-http/src/ws/dispatcher.rs
@@ -73,8 +73,8 @@ mod inner {
     use actix_service::{IntoService, Service};
     use futures_core::stream::Stream;
     use local_channel::mpsc;
-    use log::debug;
     use pin_project_lite::pin_project;
+    use tracing::debug;
 
     use actix_codec::{AsyncRead, AsyncWrite, Decoder, Encoder, Framed};