+++ actix-http/src/ws/frame.rs
@@ -1,7 +1,7 @@
 use std::convert::TryFrom;
 
 use bytes::{Buf, BufMut, BytesMut};
-use log::debug;
+use tracing::debug;
 
 use super::{
     mask::apply_mask,